LayoutTests/platform/ios-simulator/transitions/transition-hit-test-transform-expected.txt

CONSOLE MESSAGE: line 53: Transition is not running
FAIL: Timed out waiting for notifyDone to be called
This test starts a transition of the '-webkit-transform' property and then does elementFromPoint calls at the shown yellow dots to see if hit testing works
